// frontend/src/types/asset.ts
export interface Asset {
  AssetID: string;
  Picture?: string;
  MainCategory: string;
  Status: string;
  Category?: string;
  ModelName?: string;
  Brand?: string;
  Model?: string;
  CPU?: string;
  Ram?: string;
  HDD?: string;
  WLANMACAddress?: string;
  LANMACAddress?: string;
  Description?: string;
  AssignedTo?: number;
  AssignedToName?: string;
  Department?: string;
  DateAssigned?: string;
  DatePurchase?: string;
  DateFirstUse?: string;
  EndOfLife?: string;
  SerialNumber?: string;
  SNType?: string;
  PRID?: number;
  PONumber?: string;
  Price?: number;
  Invoice?: string;
  DeliveryNote?: string;
  ComputerName?: string;
  CompanyID?: number;
  CompanyName?: string;
  LocationID: number;
  LocationName?: string;
  LocationDepartment?: string;
  Accessories?: string;
  Comment?: string;
  Year?: number;
  Month?: number;
  SupplierWarrantyType?: string;
  SupplierWarrantyScope?: string;
  SupplierWarrantyServiceMethod?: string;
  SupplierWarrantyStart?: string;
  SupplierWarrantyEnd?: string;
  ReplacementCost?: number;
  AdditionalWarranty?: string;
  AdditionalWarrantyCompany?: string;
  AdditionalWarrantyType?: string;
  AdditionalWarrantyScope?: string;
  AdditionalWarrantyServiceMethod?: string;
  AdditionalWarrantyStart?: string;
  AdditionalWarrantyEnd?: string;
  TestNumber?: string;
  CreatedAt: string;
  UpdatedAt: string;
  CreatedBy?: number;
  CreatedByName?: string;
  ModifiedBy?: number;
  ModifiedByName?: string;
}